requirements.txt
wxpython==4.2.1
pyserial==3.5
numpy>=1.26.0
tornado>=6.0
pint==0.20.1
pillow==10.4.0
six==1.16.0
bitarray==3.8.0
